% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gene_list.R
\name{add_gene_list}
\alias{add_gene_list}
\title{DAVID Gene Ontology List Adder}
\usage{
add_gene_list(davidWS, DE_data, list_name)
}
\arguments{
\item{davidWS}{A DAVIDWebService class object.}

\item{DE_data}{A data frame containing 3 columns ("Gene_ID", "LFC", and "padj").}

\item{list_name}{A string that will be used to identify the submitted list.}

\item{annotation_cat}{a string or vector of strings indicating which set of GO annotation terms to use}

\item{save_file}{logical. If TRUE, files generated by the DAVID GO analysis will be saved. This value is TRUE by default.}

\item{split_at}{a character at which strings will be split. If NULL (the default), strings will not be split.}
}
\description{
DAVID Gene Ontology Wrapper
}
\examples{
add_gene_list(davidWS, DE_data, list_name)
}
