% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_IDs.R
\name{convert_IDs}
\alias{convert_IDs}
\title{ENSPANG gene ID to human gene name converter}
\usage{
convert_IDs(data_table, human_IDs)
}
\arguments{
\item{data_table}{data.fame of DEGs}

\item{human_IDs}{data.frame with human gene names}
}
\description{
converts ENSPANG gene IDs to human gene names
}
\examples{
convert_IDs()
}
