% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_Excel.R
\name{get_DEG_number}
\alias{get_DEG_number}
\title{get_DEG_number}
\usage{
get_DEG_number(expr_data)
}
\arguments{
\item{expr_data}{data.frame with at least columns 'LFC', and 'padj'}
}
\description{
Calculates the number of genes with LFC values >= 1 or <= -1 and adjusted pvalues <= 0.05
}
