% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_DEG_to_Excel.R
\name{get_term_DEGs}
\alias{get_term_DEGs}
\title{Write list of DEGs mapped to a given GO term to an Excel file}
\usage{
get_term_DEGs(GO_term, GO_dat, deg_dat, file_name, sheet_name, overwrite_file)
}
\arguments{
\item{GO_term}{a string containing the Gene Ontology term of interest}

\item{GO_dat}{data.frame with Gene Ontology analysis results to be searched for the term specified by GO_term.
Must include columns for at least 'term' and 'genes'.}

\item{deg_dat}{data.frame with columns 'Gene_ID', 'LFC', and 'padj'}

\item{file_name}{string}

\item{sheet_name}{string}

\item{overwrite_file}{logical}
}
\description{
Write list of DEGs mapped to a given GO term to an Excel file
}
\examples{
get_term_DEGs(GO_term, GO_dat, deg_dat, file_name, sheet_name, overwrite_file)
}
