% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DAVID.R
\name{run_DAVID}
\alias{run_DAVID}
\title{DAVID Gene Ontology Wrapper}
\usage{
run_DAVID(davidWS, DE_data, list_name, list_type)
}
\arguments{
\item{davidWS}{A DAVIDWebService class object.}

\item{DE_data}{A data frame containing 3 columns ('Gene_ID' or 'Transcript_ID', 'LFC', and 'padj').}

\item{list_name}{A string that will be used to identify the submitted list.}

\item{list_type}{string. Either 'gene_symbol', 'gene_ID', or 'transcript_ID'}
}
\description{
DAVID Gene Ontology Wrapper
}
\examples{
run_DAVID(davidWS, DE_data, list_name, list_type)
}
