% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DESeq2_humanOnly.R
\name{run_DESeq2_humanOnly}
\alias{run_DESeq2_humanOnly}
\title{DESeq2 wrapper}
\usage{
run_DESeq2_humanOnly(counts_data)
}
\arguments{
\item{counts_data}{A data frame containing 6 columns of raw gene or transcript counts.}
}
\description{
DESeq2 wrapper
}
\examples{
run_DESeq2_humanOnly(counts_data)
}
