% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DESeq2_noBatchCorrection.R
\name{run_DESeq2_noBatchCorrection}
\alias{run_DESeq2_noBatchCorrection}
\title{DESeq2 wrapper}
\usage{
run_DESeq2_noBatchCorrection(geneCounts, orthoList, LFC_filter)
}
\arguments{
\item{geneCounts}{A data frame containing 6 columns of raw gene counts.}

\item{orthoList}{A data frame containing two columns (Gene_ID and Human_Gene_ID)}

\item{LFC_filter}{logical. If TRUE, DEG lists are filtered by LFC values in addition to FDR values}
}
\description{
DESeq2 wrapper
}
\examples{
run_DESeq2_noBatchCorrection(geneCounts, orthoList, LFC_filter)
}
