% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DAVID.R
\name{run_FA}
\alias{run_FA}
\title{DAVID Gene Ontology functional annotation wrapper}
\usage{
run_FA(davidWS, annotation_cat, split_at, DE_data)
}
\arguments{
\item{davidWS}{A DAVIDWebService class object.}

\item{annotation_cat}{a string or vector of strings indicating which set of GO annotation terms to use}

\item{split_at}{a character at which strings will be split. If NULL (the default), strings will not be split.}

\item{DE_data}{A data frame containing 3 columns ('Gene_ID' or 'Transcript_ID', 'LFC', and 'padj').}
}
\description{
DAVID Gene Ontology functional annotation wrapper
}
\examples{
run_FA(davidWS, annotation_cat, split_at, )
}
