% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_Excel.R
\name{write_counts}
\alias{write_counts}
\title{Write count data to Excel file}
\usage{
write_counts(
  CD4T_res,
  CD8T_res,
  NK_res,
  sample_name_list,
  col_1_name,
  file_name,
  transcript_names
)
}
\arguments{
\item{CD4T_res}{data.frame with gene or transcript count data}

\item{CD8T_res}{data.frame with gene or transcript count data}

\item{NK_res}{data.frame with gene or transcript count data}

\item{sample_name_list}{list with sample names}

\item{col_1_name}{string name of column 1}

\item{file_name}{string}
}
\description{
Write count data to Excel file
}
\examples{
write_counts(CD4T_res, CD8T_res, NK_res, sample_name_list, col_1_name, file_name, transcript_names)
}
