% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DAVID.R
\name{run_DAVID}
\alias{run_DAVID}
\title{DAVID Gene Ontology Wrapper}
\usage{
run_DAVID(davidWS, DE_data, list_name, save_file = TRUE)
}
\arguments{
\item{davidWS}{A DAVIDWebService class object.}

\item{DE_data}{A data frame containing 3 columns ("Gene_ID", "LFC", and "padj").}

\item{list_name}{A string that will be used to identify the submitted list.}

\item{save_file}{logical. If TRUE, files generated by the DAVID GO analysis will be saved. This value is TRUE by default.}

\item{annotation_cat}{a string or vector of strings indicating which set of GO annotation terms to use}

\item{split_at}{a character at which strings will be split. If NULL (the default), strings will not be split.}
}
\description{
DAVID Gene Ontology Wrapper
}
\examples{
run_DAVID(davidWS, list_name)
}
