% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DESeq2.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{DESeq2 wrapper}
\usage{
run_DESeq2(geneCounts, sample_data, expt_design, coefficient, count_type)
}
\arguments{
\item{geneCounts}{data.frame containing 6 columns of raw gene or transcript counts.}

\item{sample_data}{data.frame containing experimental factrors, with the number of rows equal to the number of samples}

\item{expt_design}{a formula indicating the factors of the experimental design}

\item{coefficient}{a number indicating the coefficient to be tested}

\item{count_type}{string. Either "gene" or "transcript"}
}
\description{
DESeq2 wrapper
}
\examples{
run_DESeq2(geneCounts, sample_data, expt_design, coefficient, count_type)
}
