% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DESeq2.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{DESeq2 wrapper}
\usage{
run_DESeq2(geneCounts, orthoList)
}
\arguments{
\item{geneCounts}{A data frame containing 6 columns of raw gene counts.}

\item{orthoList}{A data frame containing two columns (Gene_ID and Human_Gene_ID)}
}
\description{
DESeq2 wrapper
}
\examples{
run_DESeq2(geneCounts, orthoList)
}
