% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edgeR.R
\name{run_edgeR}
\alias{run_edgeR}
\title{edgeR wrapper}
\usage{
run_edgeR(gene_counts)
}
\arguments{
\item{gene_counts}{A data frame containing 6 columns of raw gene counts.}

\item{ortho_list}{A data frame containing two columns (Gene_ID and Human_Gene_ID)}

\item{LFC_filter}{logical. If TRUE, DEG lists are filtered by LFC values in addition to FDR values}
}
\description{
edgeR wrapper
}
\examples{
run_edgeR(gene_counts, ortho_list)
}
