% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_Excel.R
\name{write_DEGs_to_Excel}
\alias{write_DEGs_to_Excel}
\title{Write DEG list to Excel file}
\usage{
write_DEGs_to_Excel(CD4T_res, CD8T_res, NK_res, file_name)

write_DEGs_to_Excel(CD4T_res, CD8T_res, NK_res, file_name)
}
\arguments{
\item{CD4T_res}{data.frame with columns 'Gene_ID', 'LFC', and 'padj'}

\item{CD8T_res}{data.frame with columns 'Gene_ID', 'LFC', and 'padj'}

\item{NK_res}{data.frame with columns 'Gene_ID', 'LFC', and 'padj'}

\item{file_name}{string}
}
\description{
Write DEG list to Excel file

Write DEG list to Excel file
}
\examples{
write_DEGs_to_Excel(CD4T_res, CD8T_res, NK_res, file_name)
write_DEGs_to_Excel(CD4T_res, CD8T_res, NK_res, file_name)
}
