% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_Excel.R
\name{write_DETs_to_Excel}
\alias{write_DETs_to_Excel}
\title{Write DE transcript list to Excel file}
\usage{
write_DETs_to_Excel(CD4T_res, CD8T_res, NK_res, transcript_info, file_name)
}
\arguments{
\item{CD4T_res}{data.frame with columns 'Transcript_ID', 'LFC', and 'padj'}

\item{CD8T_res}{data.frame with columns 'Transcript_ID', 'LFC', and 'padj'}

\item{NK_res}{data.frame with columns 'Transcript_ID', 'LFC', and 'padj'}

\item{transcript_info}{data.frame with at least a column 'Transcript_ID'}

\item{file_name}{string}
}
\description{
Write DEG list to Excel file
}
\examples{
write_DETs_to_Excel(CD4T_res, CD8T_res, NK_res, file_name)
}
